+++ CHANGELOG.md
@@ -404,17 +404,22 @@ Todas as mudanças neste projeto serão documentadas aqui.
 - Otimização de codigo em cima dos shuffles.  
 - Alteração na maneira que a musica é adicionada nas listas depois de um download, sendo o arquivo original e não um pseudo-arquivo.  
   
+### version 5.15.3  
+- Unificação e otimização da função de delete de musicas.  
+- Adição de um gatilho ao deletar uma musica, apagar tudo relacionado no banco.  
+- Correção no funcionamento do playPrev.  
+- Correção do botão ultima musica.  
+- Adição de up/desup na seleção de musicas.  
+- Correção de reordem nas playlists.  
+  
 ### TO FUTURE VERSIONS  
-+ Adicionar um trigger para que quando uma musica seja apagada, todas as instancias dela no banco de dados seja deletada junto.  
 + Repensar conceito de setList dinamica.
 
 + Corrigir indices não recebidos quando o item ainda for clicado manualmente no android. Verificar se não foi clicado no botão de next com o shuffle ativo, e mostrar mensagem na tela dizendo que o indice ainda não foi enviado para o desktop, alem de retornar a musica anterior. [O erro ocorre no android side, função SendMediaIndexShuffleOutOfLimits].  
 + Corrigir separação de indices não recebidos enviados pelo desktop para o android na segunda metade [O erro ocorre no android side, função SendMediaIndexShuffleOutOfLimits].  
 + Criar checkpoint para reiniciar músicas.  
 + Atualizar o README para explicar novas adições.  
-+ Fazer com que ao dar up em uma música a lista se mantenha intacta até que a faixa termine, e só depois a musica upada vai para o primeiro da lista, evitando que um up te tire do index em que estava.  
 + Criar formato de fila dinamica, para guardar musicas, organização, tipo de reprodução de stack [aleatorio/repetir], agrupamento de musicas iguais / semelhantes [slowed/spedup] TO v6.0.   
 + Criar notificador para quando uma música for upada.  
 + Corrigir inicio de musica ao recriar lista no desktop.  
-+ Corrigir Mudança de playlist antes de carregamento total deve parar de enviar a playlist antiga e começar a nova.  
 + Dividir lista de ids de musicas ja baixadas no desktop para não ocupar toda a memoria do socket.